server_backup_20250725_142708/peykan-tourism/backend/requirements.txt
# Django and DRF
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database - Using SQLite for development
# psycopg2-binary==2.9.9  # Commented out for development
dj-database-url==2.1.0

# Caching and Sessions
redis==5.0.1
django-redis==5.4.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-otp==1.3.0

# Internationalization
django-modeltranslation==0.18.12
django-parler==2.3

# Environment and Settings
python-decouple==3.8
django-environ==0.11.2

# Image Processing
Pillow>=10.2.0

# API Documentation
drf-spectacular==0.27.0

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
celery==5.3.4

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0 